Rename class AdaptiveMediaBlogsEditorConfigContributor to AMBlogsEditorConfigContributor. Rename class AdaptiveMediaBlogsEditorConfigContributorTest to AMBlogsEditorConfigContributorTest. Rename class AdaptiveMediaJournalEditorConfigContributor to AMJournalEditorConfigContributor.  Rename class AdaptiveMediaJournalEditorConfigContributorTest to AMJournalEditorConfigContributorTest.
