+++ adaptive-media-blogs-editor-configuration/src/test/java/com/liferay/adaptive/media/blogs/editor/configuration/internal/AMBlogsEditorConfigContributorTest.java
@@ -12,16 +12,16 @@
  * details.
  */
 
-package com.liferay.adaptive.media.journal.editor.configuration.internal;
+package com.liferay.adaptive.media.blogs.editor.configuration.internal;
 
 import com.liferay.adaptive.media.image.item.selector.AdaptiveMediaImageFileEntryItemSelectorReturnType;
+import com.liferay.blogs.item.selector.criterion.BlogsItemSelectorCriterion;
 import com.liferay.item.selector.ItemSelector;
 import com.liferay.item.selector.ItemSelectorCriterion;
 import com.liferay.item.selector.ItemSelectorReturnType;
 import com.liferay.item.selector.criteria.FileEntryItemSelectorReturnType;
 import com.liferay.item.selector.criteria.audio.criterion.AudioItemSelectorCriterion;
 import com.liferay.item.selector.criteria.file.criterion.FileItemSelectorCriterion;
-import com.liferay.journal.item.selector.criterion.JournalItemSelectorCriterion;
 import com.liferay.portal.json.JSONFactoryImpl;
 import com.liferay.portal.kernel.json.JSONFactoryUtil;
 import com.liferay.portal.kernel.json.JSONObject;
@@ -52,8 +52,7 @@ import org.skyscreamer.jsonassert.JSONAssert;
 /**
  * @author Sergio González
  */
-public class AdaptiveMediaJournalEditorConfigContributorTest
-	extends PowerMockito {
+public class AMBlogsEditorConfigContributorTest extends PowerMockito {
 
 	@Before
 	public void setUp() {
@@ -107,14 +106,12 @@ public class AdaptiveMediaJournalEditorConfigContributorTest
 		JSONObject jsonObject = JSONFactoryUtil.createJSONObject(
 			originalJSONObject.toJSONString());
 
-		AdaptiveMediaJournalEditorConfigContributor
-			adaptiveMediaJournalEditorConfigContributor =
-				new AdaptiveMediaJournalEditorConfigContributor();
+		AMBlogsEditorConfigContributor amBlogsEditorConfigContributor =
+			new AMBlogsEditorConfigContributor();
 
-		adaptiveMediaJournalEditorConfigContributor.setItemSelector(
-			_itemSelector);
+		amBlogsEditorConfigContributor.setItemSelector(_itemSelector);
 
-		adaptiveMediaJournalEditorConfigContributor.populateConfigJSONObject(
+		amBlogsEditorConfigContributor.populateConfigJSONObject(
 			jsonObject, _inputEditorTaglibAttributes, _themeDisplay,
 			_requestBackedPortletURLFactory);
 
@@ -164,14 +161,12 @@ public class AdaptiveMediaJournalEditorConfigContributorTest
 		JSONObject jsonObject = JSONFactoryUtil.createJSONObject(
 			originalJSONObject.toJSONString());
 
-		AdaptiveMediaJournalEditorConfigContributor
-			adaptiveMediaJournalEditorConfigContributor =
-				new AdaptiveMediaJournalEditorConfigContributor();
+		AMBlogsEditorConfigContributor amBlogsEditorConfigContributor =
+			new AMBlogsEditorConfigContributor();
 
-		adaptiveMediaJournalEditorConfigContributor.setItemSelector(
-			_itemSelector);
+		amBlogsEditorConfigContributor.setItemSelector(_itemSelector);
 
-		adaptiveMediaJournalEditorConfigContributor.populateConfigJSONObject(
+		amBlogsEditorConfigContributor.populateConfigJSONObject(
 			jsonObject, _inputEditorTaglibAttributes, _themeDisplay,
 			_requestBackedPortletURLFactory);
 
@@ -184,18 +179,17 @@ public class AdaptiveMediaJournalEditorConfigContributorTest
 	public void testAddAdaptiveMediaImageFileEntryItemSelectorReturnType()
 		throws Exception {
 
-		AdaptiveMediaJournalEditorConfigContributor
-			adaptiveMediaJournalEditorConfigContributor =
-				new AdaptiveMediaJournalEditorConfigContributor();
+		AMBlogsEditorConfigContributor amBlogsEditorConfigContributor =
+			new AMBlogsEditorConfigContributor();
 
-		JournalItemSelectorCriterion blogsItemSelectorCriterion =
-			new JournalItemSelectorCriterion();
+		BlogsItemSelectorCriterion blogsItemSelectorCriterion =
+			new BlogsItemSelectorCriterion();
 
 		blogsItemSelectorCriterion.setDesiredItemSelectorReturnTypes(
 			Collections.<ItemSelectorReturnType>singletonList(
 				new FileEntryItemSelectorReturnType()));
 
-		adaptiveMediaJournalEditorConfigContributor.
+		amBlogsEditorConfigContributor.
 			addAdaptiveMediaImageFileEntryItemSelectorReturnType(
 				blogsItemSelectorCriterion);
 
@@ -205,19 +199,12 @@ public class AdaptiveMediaJournalEditorConfigContributorTest
 		Assert.assertEquals(
 			desiredItemSelectorReturnTypes.toString(), 2,
 			desiredItemSelectorReturnTypes.size());
-
-		ItemSelectorReturnType itemSelectorReturnType0 =
-			desiredItemSelectorReturnTypes.get(0);
-
 		Assert.assertTrue(
-			itemSelectorReturnType0 instanceof
+			desiredItemSelectorReturnTypes.get(0) instanceof
 				AdaptiveMediaImageFileEntryItemSelectorReturnType);
-
-		ItemSelectorReturnType itemSelectorReturnType1 =
-			desiredItemSelectorReturnTypes.get(1);
-
 		Assert.assertTrue(
-			itemSelectorReturnType1 instanceof FileEntryItemSelectorReturnType);
+			desiredItemSelectorReturnTypes.get(1) instanceof
+				FileEntryItemSelectorReturnType);
 	}
 
 	@Test
@@ -229,11 +216,10 @@ public class AdaptiveMediaJournalEditorConfigContributorTest
 		JSONObject jsonObject = JSONFactoryUtil.createJSONObject(
 			originalJSONObject.toJSONString());
 
-		AdaptiveMediaJournalEditorConfigContributor
-			adaptiveMediaJournalEditorConfigContributor =
-				new AdaptiveMediaJournalEditorConfigContributor();
+		AMBlogsEditorConfigContributor amBlogsEditorConfigContributor =
+			new AMBlogsEditorConfigContributor();
 
-		adaptiveMediaJournalEditorConfigContributor.populateConfigJSONObject(
+		amBlogsEditorConfigContributor.populateConfigJSONObject(
 			jsonObject, _inputEditorTaglibAttributes, _themeDisplay,
 			_requestBackedPortletURLFactory);
 
@@ -253,42 +239,14 @@ public class AdaptiveMediaJournalEditorConfigContributorTest
 		JSONObject jsonObject = JSONFactoryUtil.createJSONObject(
 			originalJSONObject.toJSONString());
 
-		AdaptiveMediaJournalEditorConfigContributor
-			adaptiveMediaJournalEditorConfigContributor =
-				new AdaptiveMediaJournalEditorConfigContributor();
+		AMBlogsEditorConfigContributor amBlogsEditorConfigContributor =
+			new AMBlogsEditorConfigContributor();
 
-		adaptiveMediaJournalEditorConfigContributor.populateConfigJSONObject(
+		amBlogsEditorConfigContributor.populateConfigJSONObject(
 			jsonObject, _inputEditorTaglibAttributes, _themeDisplay,
 			_requestBackedPortletURLFactory);
 
 		JSONObject expectedJSONObject = JSONFactoryUtil.createJSONObject();
-
-		JSONAssert.assertEquals(
-			expectedJSONObject.toJSONString(), jsonObject.toJSONString(), true);
-	}
-
-	@Test
-	public void testImgIsNotAddedToAllowedContentIfEverythingWasAlreadyAllowed()
-		throws Exception {
-
-		JSONObject originalJSONObject = JSONFactoryUtil.createJSONObject();
-
-		originalJSONObject.put("allowedContent", true);
-
-		JSONObject jsonObject = JSONFactoryUtil.createJSONObject(
-			originalJSONObject.toJSONString());
-
-		AdaptiveMediaJournalEditorConfigContributor
-			adaptiveMediaJournalEditorConfigContributor =
-				new AdaptiveMediaJournalEditorConfigContributor();
-
-		adaptiveMediaJournalEditorConfigContributor.populateConfigJSONObject(
-			jsonObject, _inputEditorTaglibAttributes, _themeDisplay,
-			_requestBackedPortletURLFactory);
-
-		JSONObject expectedJSONObject = JSONFactoryUtil.createJSONObject();
-
-		expectedJSONObject.put("allowedContent", true);
 
 		JSONAssert.assertEquals(
 			expectedJSONObject.toJSONString(), jsonObject.toJSONString(), true);
@@ -306,14 +264,12 @@ public class AdaptiveMediaJournalEditorConfigContributorTest
 		JSONObject jsonObject = JSONFactoryUtil.createJSONObject(
 			originalJSONObject.toJSONString());
 
-		AdaptiveMediaJournalEditorConfigContributor
-			adaptiveMediaJournalEditorConfigContributor =
-				new AdaptiveMediaJournalEditorConfigContributor();
+		AMBlogsEditorConfigContributor amBlogsEditorConfigContributor =
+			new AMBlogsEditorConfigContributor();
 
-		adaptiveMediaJournalEditorConfigContributor.setItemSelector(
-			_itemSelector);
+		amBlogsEditorConfigContributor.setItemSelector(_itemSelector);
 
-		adaptiveMediaJournalEditorConfigContributor.populateConfigJSONObject(
+		amBlogsEditorConfigContributor.populateConfigJSONObject(
 			jsonObject, _inputEditorTaglibAttributes, _themeDisplay,
 			_requestBackedPortletURLFactory);
 
@@ -351,14 +307,12 @@ public class AdaptiveMediaJournalEditorConfigContributorTest
 		JSONObject jsonObject = JSONFactoryUtil.createJSONObject(
 			originalJSONObject.toJSONString());
 
-		AdaptiveMediaJournalEditorConfigContributor
-			adaptiveMediaJournalEditorConfigContributor =
-				new AdaptiveMediaJournalEditorConfigContributor();
+		AMBlogsEditorConfigContributor amBlogsEditorConfigContributor =
+			new AMBlogsEditorConfigContributor();
 
-		adaptiveMediaJournalEditorConfigContributor.setItemSelector(
-			_itemSelector);
+		amBlogsEditorConfigContributor.setItemSelector(_itemSelector);
 
-		adaptiveMediaJournalEditorConfigContributor.populateConfigJSONObject(
+		amBlogsEditorConfigContributor.populateConfigJSONObject(
 			jsonObject, _inputEditorTaglibAttributes, _themeDisplay,
 			_requestBackedPortletURLFactory);
 
@@ -418,14 +372,12 @@ public class AdaptiveMediaJournalEditorConfigContributorTest
 		JSONObject jsonObject = JSONFactoryUtil.createJSONObject(
 			originalJSONObject.toJSONString());
 
-		AdaptiveMediaJournalEditorConfigContributor
-			adaptiveMediaJournalEditorConfigContributor =
-				new AdaptiveMediaJournalEditorConfigContributor();
+		AMBlogsEditorConfigContributor amBlogsEditorConfigContributor =
+			new AMBlogsEditorConfigContributor();
 
-		adaptiveMediaJournalEditorConfigContributor.setItemSelector(
-			_itemSelector);
+		amBlogsEditorConfigContributor.setItemSelector(_itemSelector);
 
-		adaptiveMediaJournalEditorConfigContributor.populateConfigJSONObject(
+		amBlogsEditorConfigContributor.populateConfigJSONObject(
 			jsonObject, _inputEditorTaglibAttributes, _themeDisplay,
 			_requestBackedPortletURLFactory);
 
@@ -479,14 +431,12 @@ public class AdaptiveMediaJournalEditorConfigContributorTest
 		JSONObject jsonObject = JSONFactoryUtil.createJSONObject(
 			originalJSONObject.toJSONString());
 
-		AdaptiveMediaJournalEditorConfigContributor
-			adaptiveMediaJournalEditorConfigContributor =
-				new AdaptiveMediaJournalEditorConfigContributor();
+		AMBlogsEditorConfigContributor amBlogsEditorConfigContributor =
+			new AMBlogsEditorConfigContributor();
 
-		adaptiveMediaJournalEditorConfigContributor.setItemSelector(
-			_itemSelector);
+		amBlogsEditorConfigContributor.setItemSelector(_itemSelector);
 
-		adaptiveMediaJournalEditorConfigContributor.populateConfigJSONObject(
+		amBlogsEditorConfigContributor.populateConfigJSONObject(
 			jsonObject, _inputEditorTaglibAttributes, _themeDisplay,
 			_requestBackedPortletURLFactory);
 
@@ -518,8 +468,8 @@ public class AdaptiveMediaJournalEditorConfigContributorTest
 	private List<ItemSelectorCriterion>
 		_getBlogsItemSelectorCriterionFileEntryItemSelectorReturnType() {
 
-		JournalItemSelectorCriterion blogsItemSelectorCriterion =
-			new JournalItemSelectorCriterion();
+		BlogsItemSelectorCriterion blogsItemSelectorCriterion =
+			new BlogsItemSelectorCriterion();
 
 		blogsItemSelectorCriterion.setDesiredItemSelectorReturnTypes(
 			Collections.<ItemSelectorReturnType>singletonList(
